requirements.txt
Flask
Flask-FlatPages
Frozen-Flask
Pygments
Markdown
requests
python-dotenv
Flask-Minify
feedgen
click
beautifulsoup4
livereload
readtime==1.1.1
Pillow
